+++ libs/designer-ui/src/lib/panel/panelheader/panelheader.tsx
@@ -1,19 +1,20 @@
 import { PanelLocation, PanelScope } from '../panelUtil';
+import type { PanelNodeData } from '../types';
 import { PanelHeaderComment } from './panelheadercomment';
 import type { TitleChangeHandler } from './panelheadertitle';
 import { PanelHeaderTitle } from './panelheadertitle';
 import { Button, Menu, MenuList, MenuPopover, MenuTrigger, Spinner, Tooltip } from '@fluentui/react-components';
 import {
   bundleIcon,
-  ChevronRight24Filled,
-  ChevronRight24Regular,
+  ChevronDoubleRightFilled,
+  ChevronDoubleRightRegular,
   MoreVertical24Filled,
   MoreVertical24Regular,
+  PinOffRegular,
 } from '@fluentui/react-icons';
-import type { IButton } from '@fluentui/react/lib/Button';
 import { Icon } from '@fluentui/react/lib/Icon';
 import { css } from '@fluentui/react/lib/Utilities';
-import React, { useEffect, useMemo } from 'react';
+import { useEffect, useMemo, useRef } from 'react';
 import { useIntl } from 'react-intl';
 
 export const handleOnEscapeDown = (e: React.KeyboardEvent<HTMLInputElement | HTMLTextAreaElement>): void => {
@@ -21,80 +22,146 @@ export const handleOnEscapeDown = (e: React.KeyboardEvent<HTMLInputElement | HTM
     e.preventDefault();
   }
 };
+
 export interface PanelHeaderProps {
+  nodeData: PanelNodeData;
   isCollapsed: boolean;
+  isOutermostPanel?: boolean;
+  headerItems: JSX.Element[];
   headerLocation: PanelLocation;
-  cardIcon?: string;
-  comment?: string;
-  titleId?: string;
-  isError?: boolean;
-  isLoading?: boolean;
-  headerMenuItems: JSX.Element[];
   noNodeSelected?: boolean;
   panelScope: PanelScope;
   suppressDefaultNodeSelectFunctionality?: boolean;
   readOnlyMode?: boolean;
   renameTitleDisabled?: boolean;
-  showCommentBox?: boolean;
-  title?: string;
-  nodeId: string;
-  horizontalPadding: string;
   canResubmit?: boolean;
   resubmitOperation?: () => void;
+  onUnpinAction?: () => void;
   commentChange(panelCommentChangeEvent?: string): void;
   onRenderWarningMessage?(): JSX.Element;
   toggleCollapse: () => void;
   onTitleChange: TitleChangeHandler;
-  onTitleBlur?: (prevtitle: string) => void;
+  onTitleBlur?: (prevTitle: string) => void;
 }
 
-const DismissIcon = bundleIcon(ChevronRight24Filled, ChevronRight24Regular);
+const DismissIcon = bundleIcon(ChevronDoubleRightFilled, ChevronDoubleRightRegular);
 const OverflowIcon = bundleIcon(MoreVertical24Filled, MoreVertical24Regular);
 
-export const PanelHeader = ({
-  isCollapsed,
-  headerLocation,
-  cardIcon,
-  comment,
-  noNodeSelected,
-  isError,
-  isLoading,
-  panelScope,
-  suppressDefaultNodeSelectFunctionality,
-  titleId,
-  headerMenuItems,
-  readOnlyMode,
-  renameTitleDisabled,
-  showCommentBox,
-  title,
-  nodeId,
-  horizontalPadding,
-  canResubmit,
-  resubmitOperation,
-  commentChange,
-  onRenderWarningMessage,
-  toggleCollapse,
-  onTitleChange,
-  onTitleBlur,
-}: PanelHeaderProps): JSX.Element => {
+const CollapseButton = (props: PanelHeaderProps & { isRight: boolean; nodeId: string }): JSX.Element => {
+  const { isCollapsed, isOutermostPanel, isRight, nodeId, toggleCollapse } = props;
+
   const intl = useIntl();
 
-  const menuButtonRef = React.createRef<IButton>();
+  const menuButtonRef = useRef<HTMLButtonElement>(null);
+
+  const panelCollapseTitle = intl.formatMessage({
+    defaultMessage: 'Collapse',
+    id: 'lX30/R',
+    description: 'Text of Tooltip to collapse',
+  });
+  const panelExpandTitle = intl.formatMessage({
+    defaultMessage: 'Expand',
+    id: 'oZMhX/',
+    description: 'Text of Tooltip to expand',
+  });
+  const buttonText = isCollapsed ? panelExpandTitle : panelCollapseTitle;
+
+  const className: string = css('collapse-toggle', isRight ? 'right' : 'left', isCollapsed && 'collapsed');
+
+  useEffect(() => {
+    if (isCollapsed || !isOutermostPanel || !nodeId) {
+      return;
+    }
+
+    menuButtonRef.current?.focus();
+  }, [isCollapsed, isOutermostPanel, nodeId]);
+
+  return (
+    <Tooltip relationship="label" positioning={'before'} content={buttonText}>
+      <Button
+        id="msla-panel-header-collapse-nav"
+        ref={menuButtonRef}
+        appearance="subtle"
+        icon={<DismissIcon />}
+        className={className}
+        aria-label={buttonText}
+        onClick={toggleCollapse}
+        data-automation-id="msla-panel-header-collapse-nav"
+      />
+    </Tooltip>
+  );
+};
+
+const OverflowButton = (props: PanelHeaderProps): JSX.Element => {
+  const { headerItems } = props;
+
+  const intl = useIntl();
+
+  const panelHeaderMenuCommands = intl.formatMessage({
+    defaultMessage: 'More commands',
+    id: '0y5eia',
+    description: 'Label for commands in panel header',
+  });
+
+  return (
+    <Menu>
+      <MenuTrigger>
+        <Tooltip relationship={'label'} positioning={'before'} content={panelHeaderMenuCommands}>
+          <Button
+            appearance="subtle"
+            icon={<OverflowIcon />}
+            aria-label={panelHeaderMenuCommands}
+            data-automation-id="msla-panel-header-more-options"
+          />
+        </Tooltip>
+      </MenuTrigger>
+      <MenuPopover>
+        <MenuList>{headerItems}</MenuList>
+      </MenuPopover>
+    </Menu>
+  );
+};
+
+export const PanelHeader = (props: PanelHeaderProps): JSX.Element => {
+  const {
+    nodeData,
+    isCollapsed,
+    isOutermostPanel,
+    headerLocation,
+    noNodeSelected,
+    panelScope,
+    suppressDefaultNodeSelectFunctionality,
+    readOnlyMode,
+    renameTitleDisabled,
+    canResubmit,
+    resubmitOperation,
+    onUnpinAction,
+    commentChange,
+    onRenderWarningMessage,
+    onTitleChange,
+    onTitleBlur,
+  } = props;
+
+  const { comment, displayName: title, iconUri: cardIcon, isError, isLoading, nodeId } = nodeData;
+
+  const intl = useIntl();
 
   const resubmitButtonText = intl.formatMessage({
     defaultMessage: 'Submit from this action',
     id: 'I+85NV',
     description: 'Button label for submitting a workflow to rerun from this action',
   });
-  useEffect(() => {
-    menuButtonRef.current?.focus();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isCollapsed]);
+
+  const unpinButtonText = intl.formatMessage({
+    defaultMessage: 'Unpin action',
+    id: 'iTz1lp',
+    description: 'Text indicating a menu button to unpin a pinned action from the side panel',
+  });
 
   const isRight = headerLocation === PanelLocation.Right;
-
   const noNodeOnCardLevel = noNodeSelected && panelScope === PanelScope.CardLevel;
-  const shouldHideCollapseButton = isCollapsed && suppressDefaultNodeSelectFunctionality;
+  const shouldHideCollapseButton = !isOutermostPanel || (isCollapsed && suppressDefaultNodeSelectFunctionality);
+  const titleId = `${nodeId}-title`;
 
   // collapsed -> loading -> connector icon -> error -> backup loading
   const iconComponent = useMemo(
@@ -115,69 +182,12 @@ export const PanelHeader = ({
     [isLoading, cardIcon, isCollapsed, isError]
   );
 
-  const CollapseButton = (): JSX.Element => {
-    const panelCollapseTitle = intl.formatMessage({
-      defaultMessage: 'Collapse',
-      id: 'lX30/R',
-      description: 'Text of Tooltip to collapse',
-    });
-    const panelExpandTitle = intl.formatMessage({
-      defaultMessage: 'Expand',
-      id: 'oZMhX/',
-      description: 'Text of Tooltip to expand',
-    });
-    const buttonText = isCollapsed ? panelExpandTitle : panelCollapseTitle;
-
-    const className: string = css('collapse-toggle', isRight ? 'right' : 'left', isCollapsed && 'collapsed');
-
-    return (
-      <Tooltip relationship="label" positioning={'before'} content={buttonText}>
-        <Button
-          autoFocus={!isCollapsed}
-          id="msla-panel-header-collapse-nav"
-          appearance="subtle"
-          icon={<DismissIcon />}
-          className={className}
-          aria-label={buttonText}
-          onClick={toggleCollapse}
-          data-automation-id="msla-panel-header-collapse-nav"
-        />
-      </Tooltip>
-    );
-  };
-
-  const OverflowButton = (): JSX.Element => {
-    const PanelHeaderMenuCommands = intl.formatMessage({
-      defaultMessage: 'More commands',
-      id: '0y5eia',
-      description: 'Label for commands in panel header',
-    });
-
-    return (
-      <Menu>
-        <MenuTrigger>
-          <Tooltip relationship={'label'} positioning={'before'} content={PanelHeaderMenuCommands}>
-            <Button
-              appearance="subtle"
-              icon={<OverflowIcon />}
-              aria-label={PanelHeaderMenuCommands}
-              data-automation-id="msla-panel-header-more-options"
-            />
-          </Tooltip>
-        </MenuTrigger>
-        <MenuPopover>
-          <MenuList>{headerMenuItems}</MenuList>
-        </MenuPopover>
-      </Menu>
-    );
-  };
-
   return (
     <div className="msla-panel-header" id={noNodeOnCardLevel ? titleId : title}>
-      {shouldHideCollapseButton ? undefined : <CollapseButton />}
+      {shouldHideCollapseButton ? undefined : <CollapseButton {...props} isRight={isRight} nodeId={nodeId} />}
       {!noNodeOnCardLevel && !isCollapsed ? (
         <>
-          <div className={'msla-panel-card-header'} style={isRight ? {} : { paddingLeft: horizontalPadding }}>
+          <div className={'msla-panel-card-header'}>
             {iconComponent}
             <div className={'msla-panel-card-title-container'}>
               <PanelHeaderTitle
@@ -186,14 +196,19 @@ export const PanelHeader = ({
                 readOnlyMode={readOnlyMode}
                 renameTitleDisabled={renameTitleDisabled}
                 titleValue={title}
-                onChange={onTitleChange}
+                onChange={(newId) => onTitleChange(nodeId, newId)}
                 onBlur={onTitleBlur}
               />
             </div>
-            <OverflowButton />
+            {onUnpinAction ? (
+              <Tooltip content={unpinButtonText} relationship="label">
+                <Button appearance="subtle" icon={<PinOffRegular />} onClick={onUnpinAction} />
+              </Tooltip>
+            ) : null}
+            <OverflowButton {...props} />
           </div>
           {onRenderWarningMessage ? onRenderWarningMessage() : null}
-          {showCommentBox ? (
+          {comment ? (
             <PanelHeaderComment
               comment={comment}
               isCollapsed={isCollapsed}