+++ libs/designer/src/lib/core/state/designerView/designerViewSelectors.ts
@@ -9,6 +9,6 @@ export const useClampPan = () => {
   return useSelector((state: RootState) => state.designerView.clampPan);
 };
 
-export const useShowDeleteModal = () => {
-  return useSelector((state: RootState) => state.designerView.showDeleteModal);
+export const useShowDeleteModalNodeId = () => {
+  return useSelector((state: RootState) => state.designerView.showDeleteModalNodeId);
 };