+++ libs/designer/src/lib/ui/menuItems/pinMenuItem.tsx
@@ -1,4 +1,4 @@
-import { useIsNodePinned } from '../../core/state/panelV2/panelSelectors';
+import { useIsNodePinnedToOperationPanel } from '../../core/state/panelV2/panelSelectors';
 import { MenuItem } from '@fluentui/react-components';
 import { bundleIcon, Pin24Filled, Pin24Regular, PinOff24Filled, PinOff24Regular } from '@fluentui/react-icons';
 import { LogEntryLevel, LoggerService } from '@microsoft/logic-apps-shared';
@@ -16,19 +16,19 @@ export interface PinMenuItemProps {
 export const PinMenuItem: React.FC<PinMenuItemProps> = (props) => {
   const { nodeId, onClick } = props;
 
-  const isNodePinned = useIsNodePinned(nodeId);
+  const isNodePinned = useIsNodePinnedToOperationPanel(nodeId);
 
   const intl = useIntl();
 
   const pinAction = intl.formatMessage({
-    defaultMessage: 'Pin Action',
-    id: 'yy7jg4',
+    defaultMessage: 'Pin action',
+    id: 'b2aL+f',
     description: 'Text indicating a menu button to pin an action to the side panel',
   });
 
   const unpinAction = intl.formatMessage({
-    defaultMessage: 'Unpin Action',
-    id: 'UFMpGk',
+    defaultMessage: 'Unpin action',
+    id: 'iTz1lp',
     description: 'Text indicating a menu button to unpin a pinned action from the side panel',
   });