+++ libs/designer/src/lib/ui/panel/connectionsPanel/connectionsPanel.tsx
@@ -46,18 +46,18 @@ export const ConnectionPanel = (props: CommonPanelProps) => {
     description: 'Header for the connections panel',
   });
   const selectConnectionPanelHeader = intl.formatMessage({
-    defaultMessage: 'Change Connection',
-    id: 'cZXqqf',
+    defaultMessage: 'Change connection',
+    id: 'eb91v1',
     description: 'Header for the change connection panel',
   });
   const createConnectionPanelHeader = intl.formatMessage({
-    defaultMessage: 'Create Connection',
-    id: 'NdOhXD',
+    defaultMessage: 'Create connection',
+    id: 'NHqCeQ',
     description: 'Header for the create connection panel',
   });
   const closeButtonAriaLabel = intl.formatMessage({
-    defaultMessage: 'Close connections panel',
-    id: 'A3SsLM',
+    defaultMessage: 'Close panel',
+    id: 'uzj2d3',
     description: 'Aria label for the close button in the connections panel',
   });
 
@@ -87,7 +87,7 @@ export const ConnectionPanel = (props: CommonPanelProps) => {
     <>
       <div className="msla-app-action-header">
         <XLargeText text={panelHeaderText} />
-        <Button appearance="subtle" aria-label={closeButtonAriaLabel} onClick={props.toggleCollapse} icon={<CloseIcon />} />
+        <Button aria-label={closeButtonAriaLabel} appearance="subtle" onClick={props.toggleCollapse} icon={<CloseIcon />} />
       </div>
       <div className="msla-connections-panel-body">{renderContent()}</div>
     </>